% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_barplot.R
\name{draw_barplot}
\alias{draw_barplot}
\title{This function draws a barplot of sales volume for three years with data}
\usage{
draw_barplot(data)
}
\arguments{
\item{data}{A tibble object which has "name" & "value" columns}
}
\description{
This function draws a barplot of sales volume for three years with data
}
\examples{
\dontrun{
draw_barplot(sales)
}
}
