% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales Volume for three years}
\format{
Tibble with 799 rows and 3 columns
\describe{
\item{Company}{Company Name}
\item{Year}{Year}
\item{Sales}{Sales Volume}
}
}
\source{
\url{http://finance.naver.com}
}
\usage{
sales
}
\description{
Sales Volume of 799 Companies for three years in KOSPI
}
\examples{
  sales
}
\keyword{datasets}
