% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sales3Yrs-package.R
\docType{package}
\name{Sales3Yrs-package}
\alias{Sales3Yrs}
\alias{Sales3Yrs-package}
\title{Sales3Yrs: Barplot of Korean Company's Salse for three years}
\description{
This package has a function for drawing barplot using ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MrKevinNa/Sales3Yrs}
  \item Report bugs at \url{https://github.com/MrKevinNa/Sales3Yrs/issues}
}

}
\keyword{internal}
